refactor(web): Moved import calculations to .ts file so they can be better tested
